namespace change from `Data.Structures` to `Data.Collections`
