+++ Data_oM/Collections/Graph.cs
@@ -20,53 +20,20 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
+using System.Collections.Generic;
 
-namespace BH.oM.Data.Structures
+namespace BH.oM.Data.Collections
 {
-    public class DiscretePoint : IDataStructure, IComparable<DiscretePoint>
+    public class Graph<T> : IDataStructure
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public int X { get; set; } = 0;
-        public int Y { get; set; } = 0;
-        public int Z { get; set; } = 0;
+        public List<GraphNode<T>> Nodes { get; set; } = new List<GraphNode<T>>();
 
+        public List<GraphLink<T>> Links { get; set; } = new List<GraphLink<T>>();
 
-        /***************************************************/
-        /**** IComparable Interface                     ****/
-        /***************************************************/
-
-        public int CompareTo(DiscretePoint other)
-        {
-            if (X != other.X)
-                return X.CompareTo(other.X);
-            else if (Y != other.Y)
-                return Y.CompareTo(other.Y);
-            else
-                return Z.CompareTo(other.Z);
-        }
-
-        /***************************************************/
-
-        public override int GetHashCode()
-        {
-            return X.GetHashCode() ^ Y.GetHashCode() ^ Z.GetHashCode();
-        }
-
-        /***************************************************/
-
-        public override bool Equals(object obj)
-        {
-            if (obj is DiscretePoint)
-            {
-                DiscretePoint other = (DiscretePoint)obj;
-                return ((this.X == other.X) && (this.Y == other.Y) && (this.Z == other.Z));
-            }
-            return false;
-        }
 
         /***************************************************/
     }
